Fix license
The Free Software Foundation reached out to me and explained that I
wasn't allowed to use GPL's trademark in the way I had been using it
prior, and instructed me to follow this guide to make it compliant:

https://www.gnu.org/licenses/gpl-faq.html#ModifyGPL

Essentially, I have to change the name of the license slightly, remove
the preamble, and adjust the end section a bit. So I went ahead a fixed
that.

The terms of the license itself is unchanged, I just wasn't allowed to
call it GPL.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>